(feat) quiz-management: add publish quiz, fix bug

+++ src/main/java/com/labate/mentoringme/dto/request/quiz/FindQuizRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class FindQuizRequest {
-  private Integer userId;
-  private Integer categoryId;
+  private Long userId;
+  private Long categoryId;
 }
